refactor(session): consolidate internal session modules
