+++ scripts/generate-docs.ts
@@ -15,17 +15,14 @@ import {
   CONFIG_REFERENCE_OPTIONS,
 } from "../src/core/config";
 import { renderHunkReviewSkill } from "../src/hunk-review/skillDocument";
-import {
-  type AgentCommandOption,
-  SESSION_AGENT_COMMAND_LIST,
-} from "../src/hunk-session/agentSurface";
+import { type AgentCommandOption, SESSION_AGENT_COMMAND_LIST } from "../src/session/agent/surface";
 import {
   DEFAULT_SESSION_BROKER_HOST,
   DEFAULT_SESSION_BROKER_PORT,
   SESSION_BROKER_HOST_ENV,
   SESSION_BROKER_PORT_ENV,
   UNSAFE_ALLOW_REMOTE_SESSION_BROKER_ENV,
-} from "../src/session-broker/brokerConfig";
+} from "../src/session/broker/brokerConfig";
 
 const REPO_ROOT = resolve(import.meta.dir, "..");
 const GENERATED_NOTICE =
