+++ src/main.tsx
@@ -4,8 +4,8 @@ import { formatCliError } from "./core/errors";
 import { pagePlainText } from "./core/pager";
 import { prepareStartupPlan } from "./app/startup";
 import { sanitizeTerminalText } from "./lib/terminalText";
-import { serveSessionBrokerDaemon } from "./session-broker/brokerServer";
-import { runSessionCommand } from "./session/commands";
+import { serveSessionBrokerDaemon } from "./session/broker/brokerServer";
+import { runSessionCommand } from "./session/agent/commands";
 
 async function main() {
   const startupPlan = await prepareStartupPlan();
