+++ src/session/agent/capabilities.test.ts
