+++ src/session/agent/protocolSchemas.test.ts
