+++ src/session/broker/brokerClient.test.ts
@@ -4,8 +4,8 @@ import {
   createTestSessionRegistration,
   createTestSessionReviewFile,
   createTestSessionSnapshot,
-} from "../../test/helpers/session-daemon-fixtures";
-import { HUNK_SESSION_API_VERSION, HUNK_SESSION_DAEMON_VERSION } from "../session/protocol";
+} from "../../../test/helpers/session-daemon-fixtures";
+import { HUNK_SESSION_API_VERSION, HUNK_SESSION_DAEMON_VERSION } from "../agent/protocol";
 import { SessionBrokerClient } from "./brokerClient";
 
 const originalHost = process.env.HUNK_MCP_HOST;
